fix compile bugs in Qdraw in linux

+++ src/contracts/Qdraw.h
@@ -201,7 +201,7 @@ protected:
                 else 
                 {
                     locals.rand = qpi.K12(qpi.getPrevSpectrumDigest());
-                    locals.winner = state._participants.get(mod(locals.rand.u64._0, state._participantCount));
+                    locals.winner = state._participants.get(mod(static_cast<uint64>(locals.rand.u64._0), state._participantCount));
                     qpi.transfer(locals.winner, state._pot);
                     state._lastWinner = locals.winner;
                     state._lastWinAmount = state._pot;
